UBUNTU: Disable SUNRPC_REGISTER_V4
Bug: #306016

It seems that NFS has not been working well on Jaunty, certainly there
has been some complaining on IRC.  Kees pushed one of our bugs upstream
and they have confirmed that we should not be using CONFIG_SUNRPC_REGISTER_V4
unless we are using rpcbind over portmapper.  As we still use portmapper
we do not want this config option.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>