msm: usb: gadget: Save and restore otgsc register at h/w reset
As otgsc register is used by transceiver to read the ID pin status
and for interrupts like ID pin change and vbus(B session valid) changes.
The state of this register is cleared once the link is reset and has to
be restored to the previous state the transceiver is aware of. Hence
save the state before the reset and restore the same after reset.

Patch provided by Qualcomm

Signed-off-by: Mike Lockwood <lockwood@android.com>
Signed-off-by: Chiranjeevi Velempati <cvelempa@qualcomm.com>
(cherry picked from commit 8282fe93627113cf431a3be726eff1f74a4674b2)