[ARM] msm: i2c: Always read at least two bytes
If we tell the controller to stop at the same time as it finished reading
the first byte, it may ignore the stop request, loose bus access or fail
to start the next read or write. We now always read an extra byte if a
single byte read was requested.

Signed-off-by: Arve Hjønnevåg <arve@android.com>
Signed-off-by: David Brown <davidb@quicinc.com>