UBUNTU: ARM: IMX51: ENGR00111294 Not enter suspend state when the keypad is pressed
BugLink: https://bugs.launchpad.net/ubuntu/+bug/396601

In the function mxc_kpp_suspend, the poll_timer is deleted and the
keypad registers are not cleaned up well. So when the keypad is pressed
and the system enters the suspend state, the keypad driver can not work
well after resume. One simplest solution is to check poll_timer. If it's
still pending, don't allow the system to enter suspend.

Signed-off-by: Lily Zhang <r58066@freescale.com>
Signed-off-by: Brad Figg <brad.figg@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Amit Kucheria <amit.kucheria@canonical.com>