[ARM] msm: smem_log: Add power log debugfs node
Add a debugfs node that dumps the separate logging area used on
some modem builds to log power events. Output nothing if the power
event logging area is not present. Two nodes are present a raw node
and a symbol resolved node.

This change necessitated a minor refactoring to allow an arbitrary
number of logging areas to be added with ease.

Signed-off-by: Zach Pfeffer <zpfeffer@quicinc.com>
Signed-off-by: David Brown <davidb@quicinc.com>