UBUNTU: ARM: IMX51: ENGR00108842-1 : Enable booting with CPU frequency set in redboot.
BugLink: https://bugs.launchpad.net/ubuntu/+bug/396601

Set the CPU working points based on the frequency the CPU is booted which
can be changed using the clock command in redboot.
Also fix the CPUFREQ driver so that the BogoMips are correct when CPU freq
is set to a different value in redboot.

Signed-off-by: Ranjani Vaidyanathan-RA5478 <Ranjani.Vaidyanathan@freescale.com>
Signed-off-by: Brad Figg <brad.figg@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Amit Kucheria <amit.kucheria@canonical.com>