UBUNTU: [Config] merge kernel configs more agressively
When we build configs from the split configs for update, edit, or for a
build we merge the various component parts and then run 'make oldconfig'
over that merged configuration.  The config system understands which
options are valid and ignores any attempt to set the value of any option
which is invalid in this configuration.  This means that actually we can
safely merge options up where the other merge candidates have no value
at all.  This allows values to propogate further up the tree while
maintaining the same resulting configurations.

Modify splitconfig.pl to consider matching a value with no value as valid.
Where the value is unmergable we must temporarily mark the intermediate
levels so to prevent further merging.  Modify kernelconfig to drop these
hints in the final results.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>