[ARM] msm: clock: Add support for voting on EBI1 min rate.
There are two ways of voting on EBI1 min rate. One is the PM QoS API
(which might sleep) and the other is the ebi1_clk_set_min_rate() API
(which won't sleep). The 2nd API is needed to aggregate the acpuclock
vote and the final PM QoS vote since the acpuclock code can't afford
to sleep in the context of the suspend code (one of the callers).

Signed-off-by: Saravana Kannan <skannan@quicinc.com>
Signed-off-by: David Brown <davidb@quicinc.com>