UBUNTU: ubuntu: LIRC -- version 0.8.5
Squashed the original 4 commits into one for bisectability.

ExternalDriver: LIRC
Description: Linux Infrared Remote Control
Url: www.lirc.org
Mask:
Version: 0.8.5

This version supports compiling on 2.6.29 and 2.6.30.

Signed-off-by: Mario Limonciello <superm1@ubuntu.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>

This is the 2nd commit message:

UBUNTU: SAUCE: Fix LIRC include references for in tree build

LIRC is not upstream, and won't be going upstream in the near future.
When building in-tree, the headers are stored in a different location.

 OriginalAuthor: Mario Limonciello

Signed-off-by: Mario Limonciello <superm1@ubuntu.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>

This is the 3rd commit message:

UBUNTU: SAUCE: Don't show LIRC compilation warnings

This prevents warnings when compiling the lirc-serial
and lirc-sir drivers.

 OriginalAuthor: Mario Limonciello

Signed-off-by: Mario Limonciello <superm1@ubuntu.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>

This is the 4th commit message:

UBUNTU: [Config] Enable LIRC builds

This re-enables LIRC builds now that they can compiled against
2.6.30.

Signed-off-by: Mario Limonciello <superm1@ubuntu.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>