[ARM] msm: clock: Use system bus QoS to vote for min AXI freq.
The application processor usually sets the EBI1 clock frequency to
match the processor frequency. In some cases, a lower CPU frequency
but a higher AXI frequency may be required.

The drivers and user space applications can vote for a min EBI1 clock
frequency using the system bus QoS parameter. The QoS parameter
is to be provided in KHz. Note, that this does not guarantee a bus
bandwith. The resultant clock frequency will be atleast greater
than or equal to the QoS value.

CRs-fixed: 185976

Signed-off-by: Praveen Chidambaram <pchidamb@quicinc.com>
Signed-off-by: David Brown <davidb@quicinc.com>