UBUNTU: [Config] Dropped ubuntu/misc/wireless/acx
This driver appears to be of little use. Furthermore, according to Luis Rodriguez,
provenance of the sources may be in question.

Luis - 'I tried to review the provenance of this driver a while back, I had a
few exchanges with a few developers where the code came from. IIRC it
was from leaked sources. I stayed away and recommended everyone to
stay away.'

Signed-off-by: Tim Gardner <tim.gardner@canonical.com>