[ARM] pm: reset chip when Modem fails to respond to handshake
When Modem fails to respond to the power collapse handshake, Modem and
Apps are out of sync and Modem is in a bad state.  There is no way of
recovery other than resetting the chip.  This patch also increases the
timeout value dramatically to avoid false detection.  The wait for Modem
setting its RUN bit while Apps exiting power collapse is removed.  The
wait is not necessary and keeping it could lead to a Apps-Modem deadlock.

Signed-off-by: Ai Li <ali@quicinc.com>
Signed-off-by: David Brown <davidb@quicinc.com>