UBUNTU: ARM: IMX51: IPU irq handler deadlock fix
The babbage board boots to command prompt, after a period of inactivity,
the ipu driver turns off the display. There is a deadlock condition in
the driver when IRQs are disabled inside the IRQ handler, and this
causes the system not to respond to any keyboard inputs, patch fixes
this hang.

With this patch and the previous patch to timer, 2.6.31-rc1 will boot on
a babbage board and be usable.

Signed-off-by: Dinh Nguyen <Dinh.Nguyen@freescale.com>
Signed-off-by: Manoj Iyer <manoj.iyer@canonical.com>
Signed-off-by: Amit Kucheria <amit.kucheria@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>