UBUNTU: SAUCE: trace: add trace_event for the open() syscall
This patch uses TRACE_EVENT to add a tracepoint for the open()
syscall, in the spirit of the previous patch for sreadahead
that added the trace with ftrace.

It's not upstream because a much better system for adding tracepoints
for all syscalls is being worked on upstream, this is a stop-gap until
we have that.

Signed-off-by: Scott James Remnant <scott@ubuntu.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>