UBUNTU: updateconfigs -- alter concatenation order allow easier updates
Currently we update the configs by concatenating the config elements
config.flavour.<flavour>, config.common.<family>, and config.common.<arch>.
However this is not the natural overlay order, so if the flavour config
contains any new options which overlap generic ones they will be lost.
This makes the wholesale update of a config very difficult.

Switch the concatenation into the natural order: family, architecture,
and flavour.  This allows us to simply place a complete replacement config
in place of the flavour config and run debian/rules updateconfigs and
end up with a correctly inserted replacement config fragment.  This fix
also has the nice side effect that the config is not changed until the
operation is complete.

Signed-off-by: Andy Whitcroft <apw@canonical.com>