UBUNTU: ARM: IMX51: Add SoC and board support for Freescale mx51 platform
This patch is an extract from the patchset from Freescale to add only MX51
support.

Last patch in the series from Freescale that was merged into this patch
is 0060-ENGR00108162-2-sgtl5000-add-audio-for-imx-3stack.patch

Boards supported are:
- mx51 3stack
- babbage

Configuration tested with imx51_defconfig

We have also added a new Kconfig option - ARCH_MXC_CANONICAL under
ARCH_MXC. This Kconfig option is used to fence of Freescale code from
upstream code. The conflicting pieces of code are #ifdef'ed with
ARCH_MXC_CANONICAL so they are only compiled for our particular kernel
configuration only.

drivers/usb/*: This needs to be cleaned up further.

Signed-off-by: Amit Kucheria <amit.kucheria@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>