add arch_has_speculative_dfetch() to address dma caching issue
The Scorpion CPU speculatively reads data into the cache. This
may occur while a region of memory is being written via DMA, so
that region must be invalidated when it is brought under CPU
control after the DMA transaction finishes, assuming the DMA
was either bidirectional or from the device.

Currently both a clean and invalidate are being done for
DMA_BIDIRECTIONAL in dma_unmap_single. Only an invalidate should be
required here. There are drivers that currently rely on the clean
however so this will be removed when those drivers are updated.

CRs-fixed: 190032,186612
Signed-off-by: Steve Muckle <smuckle@quicinc.com>
Signed-off-by: David Brown <davidb@quicinc.com>