UBUNTU: ARM: IMX51: ENGR00114867 MX51: tristate i2c levelshifter
On Babbage boards there is a collision on the I2C bus between
the audio codec and the dvi port.  Reading the EDID from the
DVI leaves the I2C in a funky state, preventing the audio
driver from initializing. The fix for Babbage3.0 boards is to
control the I2C bus level shifter from a GPIO so that the level
shifter is tri-stated by default.  When reading the EDID the
level shifter is enabled briefly.

Signed-off-by: Alan Tull <r80115@freescale.com>
Signed-off-by: Amit Kucheria <amit.kucheria@canonical.com>