pmem: add more error checking in connect function.
Make sure that caller error cases like passing in file descriptor
that has the same file pointer as the passed in file pointer and
other issues (such as a null file pointer) are caught and rejected.

This fixes a recursive spinlock kernel oops that happened because
connect was invoked with the file descriptor having the same file
pointer as passed-in.

Signed-off-by: Stephen Biggs <sbiggs@quicinc.com>
Signed-off-by: David Brown <davidb@quicinc.com>