UBUNTU: [Config] enable CONFIG_AUFS_BR_RAMFS
BugLink: http://bugs.launchpad.net/bugs/414738

Enable CONFIG_AUFS_BR_RAMFS across all architectures:

    CONFIG_AUFS_BR_RAMFS=y

This option is needed to allow mounting of root correctly on diskless
installs of Mythbuntu.  This option is pretty benign simply allowing mount
of aufs branches onto the original initramfs.  This is not necessarily
safe (in that files may be hidden) if you are using pivot_root style
initrd handling which we do not.

Signed-off-by: Andy Whitcroft <apw@canonical.com>