UBUNTU: ARM: IMX51: ENGR00115688 dvi: disable edid read
Reading the EDID doesn't work for some boards/monitors/cables and
causes I2C errors that prevent sgtl5000 audio codec from initializing.

This is due to a board error in Babbage.  The decision was to disable
the I2C path to the DVI connector.  On Babbage 3.0 boards this is done
by bringing a GPIO low.  This GPIO tristates the I2C levelshifter.  The
code already exists which already does that, this ticket is to disable
code that brings the GPIO high for reading the EDID.  That way the
level shifter will remain tristated.

Signed-off-by: Alan Tull <r80115@freescale.com>
Signed-off-by: Amit Kucheria <amit.kucheria@canonical.com>