alsa: soc: 8k: Fix underrun errors in alsa concurrent playback record.
The first underrun occurs due to the volume update forcefully done
for stream even if there is no change from the default. Removed this
forceful update of volume.

Then onwards the underrun is triggered because in prepare routine
pcm_irq_pos is not reset in the second instance of the function call.
This results in a continuous mismatch in the hw/sw pointers and they
result in repeated underrun errors. Resetng the pcm_irq_pos solves the
issue.

CRs-Fixed: 192951
Signed-off-by: Asish Bhattacharya <asishb@qualcomm.com>
Signed-off-by: David Brown <davidb@quicinc.com>