pmem: add sysfs component for status of regions
NOTE: Macro usage in the code causes checkpatch to fail in two
places.  David Brown has seen this and agrees that it is a false
positive and that this will have to be manually verified and
integrated.

Add read-only elements under /sys/kernel for seeing status of PMEM
regions.

Merge debugfs functionality into the sysfs area, removing debugfs
completely.

The top directory is /sys/kernel/pmem_regions with each named PMEM
region having its own directory underneath.

In each PMEM region directory, there are files that can be read that
contain current values of various parameters and statuses for the
region desired.  The file names and types change according to which
allocator the region was defined at in the board file.

Currently, all files are read-only. There is no capability provided
as of yet to enable modification of any PMEM parameters via sysfs.

Some sysfs files are common across all regions:
base -> shows the base address of the region
size -> shows the total size of the region in bytes
allocator_type -> prints a text string defining the allocator type
mapped_regions -> displays a table of all regions mapped according
	to process id holding the mapping. This functionality had
	been exposed via debugfs.

If the allocator type of the region is ALL_OR_NOTHING, a single
sysfs file appears under the directory of that region:
allocated -> shows a 1 if this region is allocated, 0 if not

If the allocator type of the region is either BUDDY_BESTFIT or
BITMAP, two additional sysfs files appear under the directory
of that region:
total_entries -> shows the total number of quanta present in the
	region, allocated or not
quantum_size -> shows the size in bytes of the smallest allocatable
	unit

If the allocator type of the region is BUDDY_BESTFIT, a single
additional sysfs file appears under the directory of that region:
buddy_bitmap_dump -> shows a list of all indices available to the
	system and whether they are allocated or not

If the allocator type of the region is BITMAP, two additional
sysfs files appear under the directory of that region:
free_quanta -> shows the number of quanta available to be allocated
bits_allocated -> displays a table listing the current allocations
	per bit with size, index, and number of quanta allocated

So, an example sysfs hierarchy directory tree could look like this:

/sys/kernel/pmem_regions --> top level directory
| + pmem_region1 --> e.g., an ALL_OR_NOTHING allocator region
| | + base
| | + size
| | + allocator_type --> prints the string 'All or Nothing'
| | + mapped_regions
| | - allocated
| + pmem_region2 --> e.g., a BUDDY_BESTFIT allocator region
| | + base
| | + size
| | + allocator_type --> prints the string 'Buddy Bestfit'
| | + mapped_regions
| | + total_entries
| | + quantum_size
| | - buddy_bitmap_dump
| + pmem_region3 --> e.g., a BITMAP allocator region
| | + base
| | + size
| | + allocator_type --> prints the string 'Bitmap'
| | + mapped_regions
| | + total_entries
| | + quantum_size
| | + free_quanta
----- bits_allocated

Signed-off-by: Stephen Biggs <sbiggs@quicinc.com>
Signed-off-by: David Brown <davidb@quicinc.com>