msm: acpuclock: Rewrite frequency stepping
For certain configurations of SoC and modem firmware, given a starting
frequency, the list of optimal stepping frequencies to use will vary
depending on the target frequency. This is now handled by having, for
each frequency in the table, one 'up' and one 'down' entry for each PLL.

The up/down indices are also replaced with pointers to the corresponding
table entries. These pointers are calculated during boot since manual
maintenance of indices is error prone.

Signed-off-by: Matt Wagantall <mattw@quicinc.com>
Signed-off-by: David Brown <davidb@quicinc.com>