UBUNTU: [Upstream]: (drop after 2.6.31) dell-laptop: don't change softblock status if HW switch is disabled
BugLink: http://bugs.launchpad.net/bugs/418721

The "hardware" switch is tied directly to a BIOS interface that will
connect and disconnect the hardware from the bus.

If you use the software interface to request the BIOS to make these
changes, the HW switch will be in an inconsistent state and LEDs may not
reflect the state of the HW.

This patch is currently living in Andrew Morton's MM tree, but I would like
to ensure it makes Karmic.

OriginalAuthor: Mario Limonciello
OriginalLocation: http://userweb.kernel.org/~akpm/mmotm/broken-out/dell_laptop-when-the-hardware-switch-is-disabled-dont-actually-allow-changing-the-softblock-status.patch

Signed-off-by: Mario Limonciello <Mario_Limonciello@Dell.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>