UBUNTU: [Upstream]: (in 2.6.34) compal-laptop: Replace sysfs support with rfkill support
This drops the support for manually groking the files in sysfs
to turn on and off the WLAN and BT for Compal laptops in favor
of platform rfkill support.

It has been combined into a single patch to not introduce regressions
in the process of simply adding rfkill support

This patch has been submitted to LKML and acked by linux-wireless
for the rfkill changes and Alan Jenkins for other changes.

 OriginalAuthor: Mario Limonciello
 OriginalLocation: LKML
 Ignore: no

Signed-off-by: Mario Limonciello <Mario_Limonciello@Dell.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>