msm: acpuclock-8x50: Avoid configuring PLL divider mux when not using it.
Configuring the PLL divider mux unnecessarily could result in running
off of wrong PLLs for a very short duration. Doing so will cause a
crash if that PLL is not on.

One example case is when switching from running off of modem PLL to
run off of the AXI clock source.

Signed-off-by: Saravana Kannan <skannan@quicinc.com>
Signed-off-by: David Brown <davidb@quicinc.com>