msm_kgsl: Vote for maximum AXI bus speed
AXI bus doesn't always run at the highest possible frequency.
Hence vote for the highest AXI frequency when the graphics core
is in use and and remove the vote when it's not in use.
Also making sure the AXI vote is dropped during suspend state
to avoid unnecessarily running the AXI at max frequency.

Signed-off-by: Omprakash Dhyade <c_odhyad@quicinc.com>
Signed-off-by: David Brown <davidb@quicinc.com>