msm: clock: Fix AXI rate voting use of magic value
PM QOS passes only s32 integer rates through to the clock driver. The
MSM_AXI_MAX_FREQ magic value, which is used to vote for the maximum
supported AXI bus rate, is changed to respect this restriction.

The behavior and use of the magic value is also changed so that it
represents the maximium AXI rate in the correct units, regardless of
whether it's passed to ebi1_clk_set_min_rate() directly, or used
with PM QOS.

Signed-off-by: Matt Wagantall <mattw@quicinc.com>
Signed-off-by: David Brown <davidb@quicinc.com>