msm: remote-spinlock: Remove WFE instruction.
Use of the WFE instruction avoids excessive polling of memory by putting
the application processor in a wait state.  The application processor,
however, expects the remote processor to execute an SEV instruction to
cause the application processor to come out of the wait state.

The remote processors do not support the SEV instruction (older ARM
architectures, etc.) and so using the WFE instruction can lead to the
application processor waiting forever.

Hence, the WFE instruction is removed in favor of continuously polling
the remote spinlock's location in memory.

CRs-fixed: 211016, 211034
Signed-off-by: Matt Wagantall <mattw@quicinc.com>
Signed-off-by: David Brown <davidb@quicinc.com>