mmc: core: Add auto suspend functionality
Provide an interface for host drivers to register a callback function that
the core will call if the MMC bus is idle for a certain amount of time.
Host drivers can enter into low power mode upon receiving this callback.
The MMC Core takes care of resuming the host upon getting new requests from
block layer.

A sysfs entry is added for configuring the idle timeout value.  If it is
set to a -ve value, core resumes the host and never suspends it again.

A module parameter "idle_timeout" is also added to set the idle timeout
through the kernel command-line.  The default value is set to 20 sec.

Signed-off-by: Sahitya Tummala <stummala@qualcomm.com>
Signed-off-by: David Brown <davidb@quicinc.com>