msm: smd_rpcrouter: fix bug in pacmark messages sequence.
Bug in router causes packet corruption with multiple packet
message is pre-empted with a 2nd write of single packet message.

The first message (A) being a multi-packet message is started.
which contains packet A1 and A2.
A second message (B) is sent during the write sequence of message (A)
between A1 and A2 packets. In the buggy version, the packet (B1) would
use the same MID as (A1) and would cause message (A) to be
corrupted at the far end when re-assembled. Also more, the
last packet of (A2) would have a different mid, causing the far-end
to never receive the completed message. The failures caused by sending
an ill-formatted packet at the far-end are undeterministic. On the local
end, The user-space thread will be locked as no reply will be received.

CRs-fixed: 211851

Signed-off-by: Richard Ruigrok <rruigrok@quicinc.com>
Signed-off-by: David Brown <davidb@quicinc.com>