UBUNTU: SAUCE: drivers/net/fec: Convert to a platform bus driver
BugLink: http://bugs.launchpad.net/bugs/438687

Currently, the FEC device is placed on the 'virtual' bus, and so has no
driver symlink in sysfs. This symlink is required for network manager to
handle the device properly, so FEC devices are currently ignored.

This change converts the FEC driver to use the platform bus, so we get
the proper symlinks set up in sysfs. With this change, NM will bring up
the network as expected.

We're doing the platform_device_register in the driver itself, rather
than in machine-specific code. This is a bit of a hack: the driver will
always attempt to initialise a FEC even if one isn't present in the
system. However, this is no different from the existing driver.

Signed-off-by: Jeremy Kerr <jeremy.kerr@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>