UBUNTU: SAUCE: dell-laptop: Store the HW switch status internally rather than requerying every time
BugLink: http://bugs.launchpad.net/bugs/430809

If we are called from dell_rfkill_update, we can't rely on the output
from dell_send_request to infer the status of WLAN/BT/WWAN because
sometimes the BIOS sends the scancode before changing the status.

The RFKILL subsystem doesn't have a good way to "query" specifically the HW blocked bits, so we
need to store it internally instead.

We also shouldn't be querying the BIOS for the status of the softblock status because that's
supposed to be done at a higher level and if this is queried when the hardware killswitch
is in the off/disabled position, this causes the system to be stuck in a softblock state.

This patch applies on-top of Matthew Garret's two unaccepted patches that hook into the input
subsystem.

OriginalAuthor: Mario Limonciello

Signed-off-by: Mario Limonciello <superm1@ubuntu.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>