msm: remote-spinlock: Add support for DAL remote spinlocks.
DAL spinlocks operate similarly to the existing remote-spinlocks,
but are located in the DAL region of SMEM. The locks are found in
DAL chunk headers, which are identified by string names.

The remote-spinlock initialization API is modified to use strings
to specify which lock to use. Strings of the following forms are
valid:
	"D:.{1,12}"	for DAL spinlocks, where what follows the
			colon is a DAL chunk name.
	"S:[0-9]" 	for SMEM spinlocks, where what follows the
			colon is the SMEM spinlock number.

Change-Id: Iaa90af0bfee33877e64ca19dac88f4038e1971a8
Signed-off-by: Matt Wagantall <mattw@quicinc.com>
Signed-off-by: David Brown <davidb@quicinc.com>