msm: timer: use tiered method to sync DGT after power collapse
The time-sync protocol is not intended to support DGT.  When sclk wraps,
it can cause DGT to jump.  In the tiered method, GPT syncs to sclk via
the time-sync protocol, and DGT syncs to GPT.  Because both DGT and GPT
are local, syncing DGT to GPT is fast and can occur more often to detect
GPT wrapping around.

Signed-off-by: Ai Li <ali@quicinc.com>
Signed-off-by: David Brown <davidb@quicinc.com>