msm: clock: Fix proc_comm 'OFF' flags for several 7x30 clocks
Add 'OFF' flags to several clocks.  Missing 'OFF' flags for clocks
that are not enabled at boot can cause crashes when attemping to read
clock enable/disabled status (from debugfs, for example) if the clock
is not enabled first.

Unnecessary 'OFF' flags are also removed from several clocks that are
known to be on at boot.

Change-Id: I7c952414d4cac68d51d0c9f6827928f8fc20c3ad
Signed-off-by: Matt Wagantall <mattw@quicinc.com>
Signed-off-by: David Brown <davidb@quicinc.com>