ARM: Fix signal restart issues with NX and OABI compat
BugLink: http://bugs.launchpad.net/bugs/453682

The signal restarting code was placed on the user stack when OABI
compatibility is enabled.  Unfortunately, with an EABI NX executable,
this results in an attempt to run code from the non-executable stack,
which segfaults the application.

Fix this by placing the code in the vectors page, along side the
signal return code, and directing the application to that code.

Reported-by: saeed bishara <saeed.bishara@gmail.com>
Tested-by: saeed bishara <saeed.bishara@gmail.com>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
signed-off-by: Eric Miao <eric.miao@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>