UBUNTU: SAUCE: AppArmor: AppArmor wrongly reports allow perms as denied
BugLink: http://bugs.launchpad.net/bugs/453335

AppArmor file auditing wrongly reports allowed permission as being denied.
This is not a sever problem when an operation being logged contains
both an allowed and denied permission.
  eg.  /foo r, is allowed but applications tries to do open /foo for rw.
  resulting in the open failing with both rw being logged as denied.

However when quieting of known denies is being used this results in
the log containing a message that r of /foo was denied even though it
is explicitly allowed by the profile.

Fix failure auditing to only report denied permissions.

Signed-off-by: John Johansen <john.johansen@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>