UBUNTU: SAUCE: AppArmor: AppArmor disallows truncate of deleted files.
BugLink: http://bugs.launchpad.net/bugs/451375

AppArmor forces a change in application behavior that while wrong is
long standing and thus can break existing applications without any
recourse except to disable apparmor mediation of the application.

The bug/change of behavior is as follows:
The application opens a file at /some/path, it then unlinks the
file and continues to hold a reference to the file through fd.
The application then tries to truncate the file using trunc("/some/path")
instead of trunc(fd).  Since the path no longer exists and is not
visible AppArmor denies the access, but standard semantics are to
allow the truncate of the deleted file.

Change AppArmor deleted file mediation to allow allow consideration
of the deleted path name.

Signed-off-by: John Johansen <john.johansen@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>