UBUNTU: SAUCE: AppArmor: AppArmor fails to audit change_hat correctly
BugLink: http://bugs.launchpad.net/bugs/462824

When the change_hat api is used AppArmor fails to log failures and
learning events correctly.  The name being report when a hatname
is specified is incorrect, and in the complain (learning) case
the message is surpressed as the error code is reset when setting
the current hat.

This means failures are not getting properly reported, and the tools
can not be used to profile change_hat behavior.

Signed-off-by: John Johansen <john.johansen@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>