UBUNTU: SAUCE: Fix sub-flavour script to not stop on missing directories
BugLink: http://bugs.launchpad.net/bugs/453073

The way the builders run the sub-flavour script seems to include setting
an option to exit loops when an error occurs within that loop. This will
happen when a wildcard entry is not found. Fix this by catching the
error condition and while being on it also fix the name printed on
install.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Colin King <colin.king@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>