UBUNTU: SAUCE: AppArmor: Policy load and replacement can fail to alloc mem
BugLink: http://bugs.launchpad.net/bugs/458299

AppArmor dfas can under some circumstances be rather large reaching
allocations of greater than 128K for a single table.  While this is
not common this can result in memory allocation asking for page allocations
with an order of 5 or more, which will fail if memory has become fragemented.

This can cause replacement of profiles to fail.  Since profile
setup and replacement is a rare event try falling back to vmalloc if kmalloc
fails.

Signed-off-by: John Johansen <john.johansen@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>