UBUNTU: update ABI for lucid
Ignore: yes

In order to ensure we have a guarenteed upgrade path from karmic to lucid
we need to ensure that the linux-meta package will have an increasing
version from karmic to lucid and that the binary packages will have
different names.  To this end we bump the ABI leaving room for karmic to
continue to grow.  As we already are using the next block of abi numbers
for another release and to ensure we do not become confused between
releases shift ABI up by 500.

Signed-off-by: Andy Whitcroft <apw@canonical.com>