UBUNTU: SAUCE: AppArmor: Fix refcounting bug causing leak of creds
BugLink: http://bugs.launchpad.net/bugs/479115

Failure to put the cred causes a memory leak that is larger than the cred
struct, as it leaks everything it references. This happens for every
unconfined processes that does an exec, change_hat or change_profile and
passes through this function.

AppArmor when doing ptrace check for domain changes, fails to drop
the ref count on the task creds when it is unconfined.

Signed-off-by: John Johansen <john.johansen@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>