remote_spinlock: Add remote mutex support.
Remote mutex is similar to remote spinlock but allows interrupts and
other threads to preempt the current thread.

The remote mutex would be preferred over remote spinlocks in cases
where the remote processor holds the lock for long durations or when
the local processor needs to hold the lock for long durations.

Change-Id: If913cc2009758d42d4264761b647e21d0935fc48
Signed-off-by: Saravana Kannan <skannan@quicinc.com>
Signed-off-by: David Brown <davidb@quicinc.com>