Subject: Allow to bring up network interface w/o ipv6 autoconf References: 161888 When bringing up a xen bridge device, it will always be configured to use a MAC address of ff:ff:ff:ff:ff:fe.  This greatly confuses IPv6 DAD, which starts logging lots and lots of useless messages to syslog.
We really want to disable IPv6 on these interfaces, and there doesn't
seem to be a reliable way to do this without bringing the interface
up first (and triggering IPv6 autoconf). 

This patch makes autoconf (DAD and router discovery) depend on the
interface's ability to do multicast. Turning off multicast for an
interface before bringing it up will suppress autoconfiguration.

 net/ipv6/addrconf.c |    2 ++
 1 files changed, 2 insertions(+)