mmc: msm_sdcc: Fix sdcc resume sequence
A missing set of brackets creates an inadvertent logic change to the
msm_sdcc resume sequence if CONFIG_MMC_MSM7X00A_RESUME_IN_WQ is not
configured.  The redundancy in the resulting code is likely to be
optimized out by the compiler, but the source code is misleading due to
indentation.  This change inserts the missing brackets.

Signed-off-by: Brent DeGraaf <bdegraaf@quicinc.com>
Signed-off-by: David Brown <davidb@quicinc.com>