ARM: 5885/1: arm: Flush TLB entries in setup_mm_for_reboot()
BugLink: http://bugs.launchpad.net/bugs/319240
commit ad3e6c0b1fad31282a8384903ed811671d840a9c upstream

We need to do that if we tinker with the MMU entries.

This fixes the occasional bug with kexec where the new
fails to uncompress with "crc error". Most likely at
least kexec on v6 and v7 need this fix.

(bryan.wu@canonical.com: back ported it to 2.6.31 for fsl-imx51)

Signed-off-by: Tony Lindgren <tony@atomide.com>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
Signed-off-by: Bryan Wu <bryan.wu@canonical.com>