ARM: 5886/1: arm: Fix cpu_proc_fin() for proc-v7.S and make kexec work
BugLink: http://bugs.launchpad.net/bugs/319240
commit 1f667c690be3ab71036c436d8391105eee23f65b upstream

The comments in arm_machine_restart() suggest that cpu_proc_fin()
will clean and disable cache and turn off interrupts. This does
not seem to be implemented for proc-v7.S, implement it the same
way as for proc-v6.S.

This also makes kexec work for v7. Note that a related TLB and
branch traget flush patch is also needed to avoid kexec
"crc error".

Note that there are still some issues that seem to be related
to L2 cache being on and causing occasional uncompress "crc error"
with kexec. Anyways, this gets kexec mostly working on V7 for now.

(bryan.wu@canonical.com: back ported it to 2.6.31 for fsl-imx51)

Signed-off-by: Tony Lindgren <tony@atomide.com>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
Signed-off-by: Bryan Wu <bryan.wu@canonical.com>