msm: qsd8x50 platform data updates.
Bring in numerous changes to the qsd8x50 platform data.  In order
to allow the drivers themselves to be pulled in cleanly, bring
the board and device files to the latest versions.

Basicaly we're breaking the cyclic dependencies of commits by
bringing in header files first (commits before this), then the
board files themselves, and this will be followed by the drivers.

This boardfile describes devices that aren't yet committed, but
will follow in subsequent patches.

Signed-off-by: David Brown <davidb@quicinc.com>
Signed-off-by: David Brown <davidb@quicinc.com>