qup_i2c: Use DEBUG correctly as a flag and not as a value.
The DEBUG flag was being used inconsistently as value and flag,
which was causing printing of messages when not desired. This commit
takes care of that inconsistency.
Correct usage of DEBUG as flag also cleans up the code and eliminates
computation that was being done when debugging was disabled.

Change-Id: I7542b39d07fc562f375d1f977ded938bfe6fbfd4
Signed-off-by: Sagar Dharia <sdharia@quicinc.com>
(cherry picked from commit 6594dca8ecca00f8f47bb7c15a3a3b59d6d2cd8f)