msm: Add support for LID open/close detection on ST1.5
PT3661 Hall Sensor is used for sensing the open/close position
of the Lid. The msm GPIO used for detection is configured by the
gpio-keys driver. Based on the change in the status of the GPIO,
the gpio-keys driver reports the event (as SW_LID or !SW_LID)
to the input subsystem.

Signed-off-by: Anirudh Ghayal <aghayal@qualcomm.com>
(cherry picked from commit e48aabbc89cce6631a81b5be4e059fc1602498fd)