[ARM] msm: Defer partition address and size computation.
Since both 2K and 4K page devices are to be supported, the partition
information that is obtained from SMEM or ATAGs which contains the
starting block number and the number of blocks, is insufficient to
determine the partition start and end addresses. This computation is
now deferred till the Nand device type is identified during the Nand
probe phase.

Signed-off-by: Murali Palnati <palnatim@qualcomm.com>
(cherry picked from commit 414a82b92ef391368b70c6d639ef03f8a064a056)