UBUNTU: expose the kernel EXTRAVERSION in dmesg and /proc/version_signature
Our versioning has always overridden the kernel EXTRAVERSION such that
where we are based on an -rcN release candidate or have stable updates
applied that information was lost.  We now also have the issue that we
have a major drm backport in the kernel which will take stable updates
from a different source on a different release schedule and it is hard
to know which version is in the kernel.

This patch adds the kernel EXTRAVERSION to /proc/version_signature and
through that also to the kernel version banner.  This allows us to know
the exact original kernel version we are based on.  This also allows us to
augment the version information by adding additional versions to the
EXTRAVERSION.  For example the following EXTRAVERSION:

    EXTRAVERSION = .9+drm33.0

Generates the following version strings:

    $ cat /proc/version_signature
    Ubuntu 2.6.32-16.26-generic 2.6.32.9+drm33.0
    $ dmesg | grep Linux\ version
    [    0.000000] Linux version 2.6.32-16-generic (root@chloe)
	(gcc version 4.4.3 (Ubuntu 4.4.3-3ubuntu1) ) #26-Ubuntu SMP
	Fri Mar 12 18:02:3 (Ubuntu 2.6.32-16.26-generic 2.6.32.9+drm33.0)

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Amit Kucheria <amit.kucheria@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>