UBUNTU: [Config] Fix sub-flavours package conflicts
BugLink: http://bugs.launchpad.net/bugs/454827

The logic that creates debian/control information for the
virtual flavour packaging is dependent on the architecture
of the environment within which 'debian/rules clean' is run,
and as such can create an incorrect conflicts statement.

In this case it is sufficient to have both conflicts hard coded in the
same statement since the conflicting packages are architecturally
mutually exclusive.

Signed-off-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>