UBUNTU: SAUCE: AppArmor: Stop page allocation warnings that can occur on policy load
Buglink: http://launchpad.net/bugs/458299

Policy load can allocate larger than page chunks of memory, but it will
attempt to use kmalloc first, and then only fall back to vmalloc if
it needs to.  Currently the NO_WARN flag is missing from kmalloc so
it dumps a warning message, when it fails.  As this isn't a real
failure lets quiet the warnings.

Signed-off-by: John Johansen <john.johansen@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>