UBUNTU: SAUCE: AppArmor: Return string len rather than the allocation size
Buglink: http://launchpad.net/bugs/551844

AppArmor getprocattr was returning the wrong size for name for unconfined
tasks.  It returned the size of memory allocated - 1 (\0 is omitted)
instead of the size of the string.  In the case of unconfined tasks the mode
string is not output so the return size needs to be adjusted appropriately.

Signed-off-by: Kees Cook <kees.cook@canonical.com>
Signed-off-by: John Johansen <john.johansen@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>