UBUNTU: (pre-stable): input: Support Clickpad devices in ClickZone mode
Add the experimental support of new Synatpics "Clickpad" devices.

This device reports the click as the middle-button, but it doesn't set
proper capability bits.  Thus the driver needs to check the product-id
and forces to enable the button detection.

In this patch, the device behaves as "ClickZone" mode, and gives
compatible events as other normal synaptics devices so that user-space
app works as is.  In the ClickZone mode, the buttons are emulated as
clicks in the bottom button zone.  Left and right clicks are judged by
the touch position.  Clicking the narrow middle point in the button
zone gives a middle click.

Dragging can be done by keeping the button down and touching the normal
area again.  Strangely, the sequence to click after touching the area
doesn't work with this device by unknown reason...

Signed-off-by: Takashi Iwai <tiwai@suse.de>

BugLink: http://bugs.launchpad.net/bugs/516329

Acked-by: Colin King <colin.king@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Chase Douglas <chase.douglas@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>