Revert "UBUNTU: (pre-stable): input: ALPS - Add signature for HP Pavilion dm3 laptops"
This reverts commit d7dc6bc007a098ff0f842549c054002611611f29.

The commit added a signature for new ALPS touchpads. Unfortunately, the
protocol for enabling absolute events appears to have changed for this
hardware. With this commit, only strict PS/2 events are provided by the
device, meaning only relative motion events.

By reverting this commit, the ImPS/2 protocol extension is used. The
extension allows the hardware to provide scroll wheel events itself.
This behavior is preferred until the proper protocol for this device is
determined.

The upstream linux-input tree has been reverted at this time, and the
patch has been NACK'd in the -stable queue.

BugLink: http://bugs.launchpad.net/bugs/550625

Signed-off-by: Chase Douglas <chase.douglas@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>