UBUNTU: don't force module dependency checking
kernel-wedge copy-modules deals with module dependencies, and it errors
out if it can't find any interdependencies in modules.dep thinking it's
an error. However, sometimes we just want to build a kernel without any
modules. In this case, we need to override the check, which this change
does.

BugLink: http://bugs.launchpad.net/bugs/577029

Signed-off-by: Chase Douglas <chase.douglas@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>