UBUNTU: SAUCE: powerpc: fix compile error when ptrace.h is included from userspace
The commit below introduced some new userspace structures to ptrace.h
this necessarily required the inclusion of linux/types.h for the kernel
and stdlib.h for userspace:

  commit 3162d92dfb79a0b5fc03380b8819fa5f870ebf1e
  Author: Dave Kleikamp <shaggy@linux.vnet.ibm.com>
  Date:   Mon Feb 8 11:51:05 2010 +0000

    powerpc: Extended ptrace interface

However although linux/types.h is assembly safe, stdlib.h is not.  So that
when the unifdef processed versions are included in assembly files such
as occurs in eglibc, non-assembly type definitions are exposed leading
to build errors.

Looking at the original commit all of the added structures are correctly
protected by __ASSEMBLY__ wrappers, therefor the headers are actually only
required in that context.  Move the headers inside the first __ASSEMBLY__
section.

BugLink: http://bugs.launchpad.net/bugs/583733
Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>