Revert "UBUNTU: SAUCE: drm/i915: don't change DRM configuration when releasing load detect pipe"
This reverts commit 7e432ba9aae503aae4cbb7a7ed4d53e457c651e4.

This patch was originally applied for Lucid to fix some lid open/close
backlight issues. Upstream did not pull this in because they thought
there was a better solution.

I am not sure what the upstream solution is, but I had someone affected
by the bug test out the 2.6.34 mainline build. They could not reproduce
the issue, so we should revert this change to be in line with upstream.

Signed-off-by: Chase Douglas <chase.douglas@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>