UBUNTU: SAUCE: omap: remove calls to usb_nop_xceiv_register from board files
This will allow the OMAP USB drivers to be compiled in as modules. At the
moment, CONFIG_NOP_USB_XCEIV cannot be a module.

I've submitted this upstream, but the correct solution hasn't been agreed
upon. Let's use this patch for now until upstream decides how they want to
solve the problem.

Signed-off-by: Amit Kucheria <amit.kucheria@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>