UBUNTU: [Config] Enable CONFIG_RCU_FAST_NO_HZ=y
Enable CONFIG_RCU_FAST_NO_HZ=y for ports and ubuntu.  If energy
efficiency is critically important, enable this option.  It causes RCU to
attempt to accelerate grace periods in order to allow the final CPU to
enter dynticks-idle state more quickly.  Keep in mind, this does
increase the overhead of the dynticks-idle checking on systems with
large numbers of CPUs.

Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>