UBUNTU: [Config] Enable CONFIG_MMC_RICOH_MMC=y
Enable CONFIG_MMC_RICOH_MMC=y for ports and ubuntu.  This option adds a
pci quirk to disable Ricoh MMC Controller.  This proprietary controller
is unnecessary because the SDHCI driver supports MMC cards on the SD
controller, but if it is not disabled, it will steal the MMC cards away
- rendering them useless. It is safe to select this even if you don't
have a Ricoh based card reader.

Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>