UBUNTU: [Config] Enable CONFIG_SCSI_HPSA=m
Enable CONFIG_SCSI_HPSA=m for ports and ubuntu.  This option supports HP
Smart Array Controllers (circa 2009).  It is a SCSI alternative to the
cciss driver, which is a block driver.  This option allows anyone wishing
to use HP Smart Array controllers to present the devices to linux as SCSI
devices, rather than as generic block devices.

Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>