UBUNTU: ubuntu: AUFS -- track changes to the arguements to fop fsync()
The file operation fsync dropped an unused struct dentry parameter,
apply this change to aufs2:

  commit 7ea8085910ef3dd4f3cad6845aaa2b580d39b115
  Author: Christoph Hellwig <hch@lst.de>
  Date:   Wed May 26 17:53:25 2010 +0200

    drop unused dentry argument to ->fsync

WARNING: this currently removes support for syncing some forms of aufs
directory as we are now required to have a filp to sync against and aufs
does not always have them.  I have added a WARN_ON to try and track the
triggers for this mode if any.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>