ptrace: limit scope to attach only (allow read)
The modification to may_ptrace_access is to broad, and would block
access to things in /proc/$pid/ that many applications expect to be
able to see.  Since the goal is to stop arbitrary code injection,
allow PTRACE_MODE_READ but not PTRACE_MODE_ATTACH for non-children.

BugLink: https://launchpad.net/bugs/589656

Signed-off-by: Kees Cook <kees.cook@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>