UBUNTU: [Upstream] (evdev) Convert to dynamic event buffer (rev4)
Allocate the event buffer dynamically, and prepare to compute the
buffer size in a separate function. This patch defines the size
computation to be identical to the current code, and does not contain
any logical changes.

This is expected to land in 2.6.36 and should be applied to the
upstream input tree momentarily:

http://git.kernel.org/?p=linux/kernel/git/dtor/input.git;a=summary

Signed-off-by: Henrik Rydberg <rydberg@euromail.se>
Acked-by: Chase Douglas <chase.douglas@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>