UBUNTU: [Upstream] (evdev) Use driver hint to compute the evdev buffer size (rev2)
Some devices, in particular MT devices, produce a lot of data.  This
leads to a high frequency of lost packets in evdev, which by default
uses a fairly small event buffer. Let the drivers hint the average
number of events per packet for the device by calling the
input_set_events_per_packet(), and use that information when computing
the evdev buffer size.

This is expected to land in 2.6.36 and should be applied to the
upstream input tree momentarily:

http://git.kernel.org/?p=linux/kernel/git/dtor/input.git;a=summary

Signed-off-by: Henrik Rydberg <rydberg@euromail.se>
Acked-by: Chase Douglas <chase.douglas@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>