Revert "UBUNTU: SAUCE: toshiba_acpi -- pull in current -dev version of driver"
This reverts commit 01f5cc3a18524b43f5385960bccb8eeabe191b3f.

Most of that original commit has now been activly upstreamed
and the driver itself is being maintained upstream.  The remaining bits
only create the older /proc interfaces which if memory serves was
only of interest to HAL which is now gone.  I therefore suggest we drop
this patch and see if anyone notices.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>