Revert "UBUNTU: SAUCE: pm: Config option to disable handling of console during suspend/resume"
This reverts commit 6063b4dc3721a63d70f81522fa130372ded60b45.

BugLink: https://bugs.launchpad.net/bugs/594885

This patch is causing suspend/resume issues in the latest Maverick
2.6.35 kernels.  It was narrowed down to the CONFIG_PM_DISABLE_CONSOLE
being enabled.  This config option originated with this SAUCE patch
which has stagnated over the years.  It's been confirmed it is no longer
necessary and just adds bloat to our distro kernel so drop it.

https://lists.ubuntu.com/archives/kernel-team/2010-June/011166.html

Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>