UBUNTU: [Config] enable CONFIG_IPV6_SIT_6RD
Enable IPv6 Rapid Deployment tunnelling support.  Whilst this support
is still marked EXPERIMENTAL this support is going to be a corner stone
of IPv6 deployments in existing IPv4 ISPs.  Pressure is being placed on
ISPs to convert over to IPv6 due to IPv4 address scarcity.  Looking at the
code this enabled an extension to the current 6to4 tunnelling mechanism.
Both 6to4 tunnelling and 6rd are opt-in mechanisms.

BugLink: http://bugs.launchpad.net/bugs/591869
Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>