UBUNTU: [Config] Disable CONFIG_DRM_VMWGFX (staging driver)
BugLink: http://bugs.launchpad.net/bugs/606139

Per a request form the VMware QA team, disable the vmwgfx driver in
staging.  It's been noted that the vmwgfx driver in staging is not
stable enough for use at the moment.  It's recommended this driver only
be enabled when it has been promoted from staging into the official
mainline tree.

Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>